Move Spinner to components package
