+++ src/attachment/AttachmentLoader.tsx
@@ -9,7 +9,7 @@
 import React from 'react';
 
 import PdfJs from '../PdfJs';
-import Spinner from '../Spinner';
+import Spinner from '../components/Spinner';
 import AttachmentList from './AttachmentList';
 import FileItem from './FileItem';
