+++ src/components/Spinner.tsx
@@ -8,20 +8,20 @@
 
 import React from 'react';
 
+import ThemeContent from '../theme/ThemeContext';
 import './spinner.less';
 
 const Spinner: React.FC<{}> = () => {
+    const theme = React.useContext(ThemeContent);
+
     return (
-        <svg className="viewer-spinner" width="64px" height="64px" viewBox="0 0 32 32">
+        <svg className={`${theme.prefixClass}-spinner`} width="64px" height="64px" viewBox="0 0 32 32">
             <circle
+                className={`${theme.prefixClass}-spinner-circle`}
                 cx="16"
                 cy="16"
-                fill="none"
                 r="12"
-                stroke="rgba(0, 0, 0, 0.4)"
                 strokeDasharray={Math.PI * 2 * 9}
-                strokeLinecap="round"
-                strokeWidth="2"
             />
         </svg>
     );
