+++ src/layers/PageLayer.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import Observer, { VisibilityChanged } from '../Observer';
 import PdfJs from '../PdfJs';
 import Match from '../search/Match';
-import Spinner from '../Spinner';
+import Spinner from '../components/Spinner';
 import ThemeContent from '../theme/ThemeContext';
 import { SpecialLevel } from '../zoom/zoomingLevel';
 import AnnotationLayer from './AnnotationLayer';
