+++ src/loader/DocumentLoader.tsx
@@ -9,7 +9,7 @@
 import React from 'react';
 
 import PdfJs from '../PdfJs';
-import Spinner from '../Spinner';
+import Spinner from '../components/Spinner';
 import ThemeContent from '../theme/ThemeContext';
 import AskForPasswordState from './AskForPasswordState';
 import AskingPassword from './AskingPassword';
